requirements.txt
slackclient==1.0.6
mysql-connector==2.1.4
